+++ src/infobox/toLatex.js
